requirements.txt
streamlit
numpy
scikit-learn
onnx
onnxruntime
openai
python-dotenv
skl2onnx